Function Related to Calculating Phase

Scalar Channel ($\sigma_{1,2}$)

PhaseGN.realpart_σ — Function

realpart_σ(temp, μ, ω,param)

Returns the real part of the polarisation for σ₁ and σ₂ at zero external momentum.

PhaseGN.imagpart_σ — Function

imagpart_σ(temp, μ, ω,param)

Returns the imaginary part of the polarisation for σ₁ and σ₂ at zero external momentum.

PhaseGN.Π0_σ — Function

Π0_σ(temp,μ,param)

Returns the momentum and frequency independent part of the polarisation function

PhaseGN.phasesc_σ — Function

phasesc_σ(temp,μ,ω,param)

Returns the scattering part of the phase for σ₁ and σ₂ at zero external momentum.

Also consider the equivalent function

phasesc_ϕ(temp,μ,ω,m,param)

where you supply the values of $\bar{\sigma}_1 = m$. Since these values only depend on the temp and μ, this function is more efficient if you want to calculate phases at different vaules of frequencies at a fixed temp and μ.

PhaseGN.phaser_σ — Function

phaser_σ(temp,μ,ω,param)

Returns the resonant part of the phase for σ₁ and σ₂ at zero external momentum.

Also consider the equivalent function

phaser_ϕ(temp,μ,ω,m,Π00,param)

where you supply the values of $\bar{\sigma}_1 = m$ and $\Pi 00$ the frequency and momentum independent part of the polarisation. Since these values only depend on the temp and μ, this function is more efficient if you want to calculate phases at different vaules of frequencies at a fixed temp and μ.

Missing docstring.

Missing docstring for phase_σ. Check Documenter's build log for details.

Pseudo-Scalar Channel ($\phi_{1,2}$)

Missing docstring.

Missing docstring for realpart_ϕ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for imagpart_ϕ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PhaseGN.Π0_ϕ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for phasesc_ϕ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for phaser_ϕ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for phase_ϕ. Check Documenter's build log for details.